Update server code and stablish a new structure for the project
