--- index.js
@@ -1,23 +1,0 @@
-'use strict'
-
-// const debug = require('debug')('nodestr:server')
-// const express = require('express')
-// const app = express()
-// app.set('port', port)
-
-const http = require('http')
-const hostname = '127.0.0.1'
-const port = 3000
-const server = http.createServer((req, res) => {
-    res.statusCode = 200
-    res.setHeader('Content-Type', 'text/plain')
-    res.end('Hello there!\n')
-})
-
-// const router = express.Router()
-
-// app.use('/', route)
-
-server.listen(port, hostname, () => {
-    console.log(`Server running at http://${hostname}:${port}`)
-})
